FastAPI And React/requirements.txt
SQLAlchemy
psycopg2-binary
email-validator
fastapi-mail
python-dotenv
--upgrade certifi